Match babycloudrunner sandworm_turnTowardTargetAnim t conversion
Use int t + (f32)(s16)t signed conversion path to drop the spurious
per-store extsh and select the named f64 magic (lbl_803E4220) instead
of the @129 unsigned-bias label. 90.9 -> 94.1%.

+++ src/main/dll/dll_014C_babycloudrunner.c
@@ -694,7 +694,7 @@ void sandworm_turnTowardTargetAnim(int a, int b, u8* c, int d)
         }
         else
         {
-            s16 t;
+            int t;
             if ((int)(s16)shifted > 0)
             {
                 t = (s16)shifted >> 2;
@@ -703,7 +703,7 @@ void sandworm_turnTowardTargetAnim(int a, int b, u8* c, int d)
             {
                 t = -(s16)shifted >> 2;
             }
-            ((ObjAnimAdvanceObjectFirstF32Fn)ObjAnim_AdvanceCurrentMove)(a, (f32)t / lbl_803E4240, timeDelta, 0);
+            ((ObjAnimAdvanceObjectFirstF32Fn)ObjAnim_AdvanceCurrentMove)(a, (f32)(s16)t / lbl_803E4240, timeDelta, 0);
         }
     }
 }
